Move files with path changes only
